support-sequence
Added support for the control sequences.
The implementation of auto-increment columns in two ways: through the
automatic generation of sequences and before insert trigger, or using
identity fields (only in Firebird 3.0).
The implementation of InsertGetId method similar to the postgres, ie
using RETURNING proposal. For Firebird 3.0 found a workaround.
Create your own base model class in which insertAndSetId method is
implemented through the prior receipt by the sequence identifier.
Added additional methods for the execution of stored procedures and
stored functions.
Added Providing the connection parameters: the name of the role and
Firebird version (to use the correct grammar).